+++ src/models.rs
@@ -7,7 +7,7 @@ use super::schema::users;
 #[diesel(table_name = crate::schema::users)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Users {
-    pub user_id: i32,
+    pub user_id: i64,
     pub username: String,
     pub email: String,
     pub pw_hash: String,
@@ -25,33 +25,33 @@ pub struct NewUser<'a> {
 #[diesel(table_name = crate::schema::followers)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Followers {
-    pub who_id: i32,
-    pub whom_id: i32,
+    pub who_id: i64,
+    pub whom_id: i64,
 }
 
 #[derive(Insertable)]
 #[diesel(table_name = followers)]
 pub struct NewFollower<'a> {
-    pub who_id: &'a i32,
-    pub whom_id: &'a i32,
+    pub who_id: &'a i64,
+    pub whom_id: &'a i64,
 }
 
 #[derive(Queryable, Selectable, Insertable)]
 #[diesel(table_name = crate::schema::messages)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Messages {
-    pub message_id: i32,
-    pub author_id: i32,
+    pub message_id: i64,
+    pub author_id: i64,
     pub text: String,
     pub pub_date: String,
-    pub flagged: i32,
+    pub flagged: i64,
 }
 
 #[derive(Insertable)]
 #[diesel(table_name = messages)]
 pub struct NewMessage<'a> {
-    pub author_id: &'a i32,
+    pub author_id: &'a i64,
     pub text: &'a str,
     pub pub_date: &'a String,
-    pub flagged: &'a i32,
+    pub flagged: &'a i64,
 }
